server/api/workers/AECIF_NET/requirements.txt
scipy
numpy
matplotlib
opencv-python-headless
torch
torchvision
tqdm
Pillow
h5py
celery
requests
